feat: 재료 페이지 클라이언트 컴포넌트 및 관련 훅 추가
- `IngredientsPage` 컴포넌트를 클라이언트 전용 `IngredientsPageClient`로 분리하여 코드의 가독성과 유지보수성을 향상시켰습니다.
- 재료 목록을 무한 스크롤로 가져오는 `useInfiniteIngredients` 훅을 추가하여 데이터 페칭 로직을 개선하였습니다.
- 그리드 애니메이션을 처리하는 `useGridAnimation` 훅을 추가하여 UI 애니메이션을 관리하도록 하였습니다.
- 재료 관리 상태를 처리하는 `useIngredientsManager` 훅을 추가하여 상태 관리를 간소화하였습니다.
