+++ src/widgets/IngredientsPage/hooks/useGridAnimation.ts
@@ -1,0 +1,64 @@
+import { useEffect, useRef } from "react";
+
+import gsap from "gsap";
+
+type UseGridAnimationProps = {
+  ingredients?: any[];
+  error?: Error | null;
+};
+
+export const useGridAnimation = ({ ingredients, error }: UseGridAnimationProps) => {
+  const gridItemsContainerRef = useRef<HTMLDivElement | null>(null);
+  const gridAnimateTargetRef = useRef<HTMLDivElement | null>(null);
+
+  useEffect(() => {
+    if (!ingredients || ingredients.length === 0 || error) {
+      return;
+    }
+
+    const ctx = gsap.context(() => {
+      if (gridAnimateTargetRef.current) {
+        const newItems = Array.from(
+          gridAnimateTargetRef.current.querySelectorAll<HTMLElement>(
+            ':scope > *:not([data-gsap-animated="true"])'
+          )
+        );
+
+        if (newItems.length > 0) {
+          gsap.fromTo(
+            newItems,
+            { opacity: 0, y: 30, scale: 0.98 },
+            {
+              opacity: 1,
+              y: 0,
+              scale: 1,
+              duration: 0.5,
+              stagger: {
+                each: 0.08,
+                onStart: function () {},
+              },
+              ease: "power2.out",
+              scrollTrigger: {
+                trigger: gridAnimateTargetRef.current,
+                start: "top 85%",
+                toggleActions: "play none none none",
+              },
+              onComplete: () => {
+                newItems.forEach((item) => {
+                  item.setAttribute("data-gsap-animated", "true");
+                });
+              },
+            }
+          );
+        }
+      }
+    }, gridItemsContainerRef);
+
+    return () => ctx.revert();
+  }, [ingredients, error]);
+
+  return {
+    gridItemsContainerRef,
+    gridAnimateTargetRef,
+  };
+};
